% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskregions.r
\name{riskregions}
\alias{riskregions}
\title{Loads IETC project risk regions as an sf object in WGS84}
\usage{
riskregions(data)
}
\arguments{
\item{data}{Accepts a string of either "USFE" or "SFB"}
}
\description{
Loads IETC project risk regions as an sf object in WGS84
}
\examples{
riskregions <- riskregions("USFE")
}
