% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzipShape.r
\name{unzipShape}
\alias{unzipShape}
\title{IETC}
\usage{
unzipShape(InputZipShapefile)
}
\arguments{
\item{InputZipShapefile}{Web address of a web-accessible zippped shapefile}
}
\value{
sf object loaded from input shapefile
}
\description{
Unzip zipped shapefiles from a download web address such as a raw GitHub link
}
